HTML/requirements.txt
bottle
mysql-connector-python
redis
bottle-session
bottle-redis
python-dotenv
PyJWT